Move children to branch patterns
