+++ tests/DocoptNet.Tests/PatternFixRepeatingArgumentsTests.cs
@@ -6,6 +6,7 @@ namespace DocoptNet.Tests
     [TestFixture]
     public class PatternFixRepeatingArgumentsTests
     {
+        /* FIXME
         [Test]
         public void Should_fix_option()
         {
@@ -23,7 +24,7 @@ namespace DocoptNet.Tests
                 new Argument("A").FixRepeatingArguments()
                 );
         }
-
+        */
         [Test]
         public void Should_fix_required_args()
         {
